Follow pixel-graphics' PaletteIndex/isize-position API changes
Adjusts call sites for pixel-graphics' newtyped PaletteIndex (Paint::Solid
now takes one instead of a raw palette index) and its isize-position API
(Framebuffer::set_pixel/row/fill_rect_paint, Rect::new, font draw calls).
splitwm's own window-geometry model (LeafView, tree, arrange/event
handling) stays i32 throughout, matching X11's native types — only the
values crossing into pixel-graphics/pixel-fonts get cast at that boundary,
the same way X11's i16 event coordinates already were.

Also drops icon.rs's own copy of magick_decode_rgba/parse_pam_rgba now
that pixel-graphics exposes it directly, and the default-features/baked
feature flags now that baking is just how those crates work.

+++ src/render/notify_popup.rs
@@ -106,7 +106,7 @@ impl Renderer {
                     );
                 let idx = bubble.at(sx, sy);
                 if idx != pixel_graphics::TRANSPARENT {
-                    fb.set_pixel(dx, dy, idx);
+                    fb.set_pixel(dx as isize, dy as isize, idx);
                 }
             }
         }
@@ -114,10 +114,22 @@ impl Renderer {
         if let Some(font) = &self.font {
             let mut y = NOTE_PAD_TOP;
             for (line, bold) in &lines {
-                font.draw_text(&mut fb, line, NOTE_PAD_LEFT, y, palette_color::BLACK);
+                font.draw_text(
+                    &mut fb,
+                    line,
+                    NOTE_PAD_LEFT as isize,
+                    y as isize,
+                    palette_color::BLACK,
+                );
                 if *bold {
                     // Faux bold: restrike one pixel right.
-                    font.draw_text(&mut fb, line, NOTE_PAD_LEFT + 1, y, palette_color::BLACK);
+                    font.draw_text(
+                        &mut fb,
+                        line,
+                        (NOTE_PAD_LEFT + 1) as isize,
+                        y as isize,
+                        palette_color::BLACK,
+                    );
                 }
                 y += line_h;
             }
